+++ pingoo/tls/mod.rs
@@ -1,5 +1,5 @@
 mod tls_manager;
-pub use tls_manager::TlsManager;
+pub use tls_manager::*;
 
 pub mod acme;
 pub mod certificate;